refactor(swift): split scan responsibilities
